third_party/blink/web_tests/platform/generic/fast/dom/XMLHttpRequest-legacy-event-listener-expected.txt

Test addEventListener() and removeEventListener() fail silently if arguments are missing.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS new XMLHttpRequest().addEventListener("foo") threw exception TypeError: Failed to execute 'addEventListener' on 'EventTarget': 2 arguments required, but only 1 present..
PASS new XMLHttpRequest().removeEventListener("bar") threw exception TypeError: Failed to execute 'removeEventListener' on 'EventTarget': 2 arguments required, but only 1 present..
PASS new XMLHttpRequest().addEventListener() threw exception TypeError: Failed to execute 'addEventListener' on 'EventTarget': 2 arguments required, but only 0 present..
PASS new XMLHttpRequest().removeEventListener() threw exception TypeError: Failed to execute 'removeEventListener' on 'EventTarget': 2 arguments required, but only 0 present..
PASS successfullyParsed is true

TEST COMPLETE
